pyproject.toml
[project]
name = "anamnesis"
version = "0.1.0"
description = "Knowledge management framework for LLM agent systems"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
api = ["fastapi>=0.115", "uvicorn>=0.34"]
dev = ["pytest>=8.0", "ruff>=0.11"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anamnesis"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 88
